feat(components): restructure component paths to use shadcn directory for better organization and clarity feat(ui): implement new UI components including Button, Input, Separator, and more for enhanced user interface feat(sidebar): create Sidebar component with context management for state handling and mobile responsiveness feat(hooks): add useIsMobile hook to determine mobile view for responsive design feat(breadcrumb): introduce Breadcrumb component for improved navigation structure feat(skeleton): add Skeleton component for loading states in UI feat(tooltip): implement Tooltip component for enhanced user interaction feedback
