+++ client/src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import {AppSidebar} from '@/components/app-sidebar';
+import {AppSidebar} from '@/components/shadcn/app-sidebar.tsx';
 import {
     Breadcrumb,
     BreadcrumbItem,
@@ -6,13 +6,13 @@ import {
     BreadcrumbList,
     BreadcrumbPage,
     BreadcrumbSeparator
-} from '@/components/ui/breadcrumb';
-import {Separator} from '@/components/ui/separator';
+} from '@/components/shadcn/ui/breadcrumb';
+import {Separator} from '@/components/shadcn/ui/separator';
 import {
     SidebarInset,
     SidebarProvider,
     SidebarTrigger
-} from '@/components/ui/sidebar';
+} from '@/components/shadcn/ui/sidebar';
 
 export function Page() {
     return (
