+++ client/src/components/shadcn/ui/collapsible.tsx
