+++ client/src/components/shadcn/ui/sheet.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import {cn} from '@/lib/utils';
+import {cn} from '@/components/shadcn/lib/utils.ts';
 import * as SheetPrimitive from '@radix-ui/react-dialog';
-import {type VariantProps, cva} from 'class-variance-authority';
+import {cva, type VariantProps} from 'class-variance-authority';
 import {X} from 'lucide-react';
 import * as React from 'react';
