+++ client/src/components/shadcn/ui/sidebar.tsx
@@ -1,17 +1,16 @@
-import {Button} from '@/components/ui/button';
-import {Input} from '@/components/ui/input';
-import {Separator} from '@/components/ui/separator';
-import {Sheet, SheetContent} from '@/components/ui/sheet';
-import {Skeleton} from '@/components/ui/skeleton';
+import {useIsMobile} from '@/components/shadcn/hooks/use-mobile.tsx';
+import {cn} from '@/components/shadcn/lib/utils.ts';
+import {Button} from '@/components/shadcn/ui/button.tsx';
+import {Input} from '@/components/shadcn/ui/input.tsx';
+import {Separator} from '@/components/shadcn/ui/separator.tsx';
+import {Sheet, SheetContent} from '@/components/shadcn/ui/sheet.tsx';
+import {Skeleton} from '@/components/shadcn/ui/skeleton.tsx';
 import {
     Tooltip,
     TooltipContent,
     TooltipProvider,
     TooltipTrigger
-} from '@/components/ui/tooltip';
-
-import {useIsMobile} from '@/hooks/use-mobile';
-import {cn} from '@/lib/utils';
+} from '@/components/shadcn/ui/tooltip.tsx';
 import {Slot} from '@radix-ui/react-slot';
 import {cva, VariantProps} from 'class-variance-authority';
 import {PanelLeft} from 'lucide-react';
