+++ client/src/components/shadcn/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import {cn} from '@/lib/utils';
+import {cn} from '@/components/shadcn/lib/utils.ts';
 
 function Skeleton({className, ...props}: React.HTMLAttributes<HTMLDivElement>) {
     return (
