+++ client/src/components/shadcn/ui/tooltip.tsx
@@ -1,4 +1,4 @@
-import {cn} from '@/lib/utils';
+import {cn} from '@/components/shadcn/lib/utils.ts';
 import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 import * as React from 'react';
